script for creating base image based on an exsiting guest
Takes image of an existing machine and creates a base image from it. Then it
creates new copy-on-write image based on this base image and replaces the
original file (to save disk space).

It's useful for creating a base image from a machine that was already created
using the snap-guest tool. The resulting base image is not copy-on-write any
more and it's independent from the original base.